+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **kiarina-llm**: Refactored internal module structure and added default values to RunContext fields
+
 ## [1.15.0] - 2025-12-13
 
 ### Added